assets/sdp-checkpoint-recovery/template/{{.target_dir}}/requirements.txt
databricks-sdk>=0.100.0
